ai/requirements.txt
Flask==2.3.2
pyodbc==4.0.35
pandas==2.0.3
openai==0.27.8
azure-ai-textanalytics==5.3.0
azure-core==1.27.0
sqlalchemy==2.0.20